Rename x ansatz following public/private API convention
