+++ src/python/zquantum/qaoa/ansatzes/__init__.py
@@ -1,3 +1,3 @@
-from .farhi_ansatz import QAOAFarhiAnsatz
+from .farhi_ansatz import QAOAFarhiAnsatz, create_farhi_qaoa_circuits
 from .warm_start_ansatz import WarmStartQAOAAnsatz
-from .x_ansatz import XAnsatz, XZAnsatz
+from ._x_ansatz import XAnsatz, XZAnsatz
