+++ src/python/zquantum/qaoa/ansatzes/_x_ansatz.py
